# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Patient Tracker App",
        "contact": {
            "name": "Mahmoud Mohamed",
            "url": "https://www.linkedin.com/in/mahmoud-mohamed-955932b5/",
            "email": "m.mohamed@cat.com.eg"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/targets": {
            "get": {
                "tags": [
                    "targets"
                ],
                "summary": "get all targets paginated",
                "parameters": [
                    {
                        "name": "hospital_id",
                        "in": "query",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "targets"
                ],
                "summary": "add add new target",
                "parameters": [
                    {
                        "name": "number",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "year",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "month",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "hospital_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/targets/{target}": {
            "put": {
                "tags": [
                    "targets"
                ],
                "summary": "update target",
                "parameters": [
                    {
                        "name": "target",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "number",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "year",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "month",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "hospital_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "_method:put",
                        "in": "formData",
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "targets"
                ],
                "summary": "Delete target",
                "parameters": [
                    {
                        "name": "target",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        }
    },
...
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "Value: Bearer \\<token\\><br><br>",
            "name": "Authorization",
            "in": "header"
        }
    }
}